Move more code to AbstractAggregate
