+++ h2/src/main/org/h2/expression/aggregate/AbstractAggregate.java
@@ -5,6 +5,9 @@
  */
 package org.h2.expression.aggregate;
 
+import org.h2.command.dml.Select;
+import org.h2.command.dml.SelectGroups;
+import org.h2.engine.Session;
 import org.h2.expression.Expression;
 import org.h2.table.ColumnResolver;
 import org.h2.table.TableFilter;
@@ -14,9 +17,20 @@ import org.h2.table.TableFilter;
  */
 public abstract class AbstractAggregate extends Expression {
 
+    protected final Select select;
+
+    protected final boolean distinct;
+
     protected Expression filterCondition;
 
     protected Window over;
+
+    private int lastGroupRowId;
+
+    AbstractAggregate(Select select, boolean distinct) {
+        this.select = select;
+        this.distinct = distinct;
+    }
 
     /**
      * Sets the FILTER condition.
@@ -57,6 +71,67 @@ public abstract class AbstractAggregate extends Expression {
             over.setEvaluatable(tableFilter, b);
         }
     }
+
+    @Override
+    public void updateAggregate(Session session, boolean window) {
+        if (window != (over != null)) {
+            if (!window && select.isWindowQuery()) {
+                updateGroupAggregates(session);
+                if (filterCondition != null) {
+                    filterCondition.updateAggregate(session, false);
+                }
+                over.updateAggregate(session, false);
+            }
+            return;
+        }
+        // TODO aggregates: check nested MIN(MAX(ID)) and so on
+        // if (on != null) {
+        // on.updateAggregate();
+        // }
+        SelectGroups groupData = select.getGroupDataIfCurrent(window);
+        if (groupData == null) {
+            // this is a different level (the enclosing query)
+            return;
+        }
+
+        int groupRowId = groupData.getCurrentGroupRowId();
+        if (lastGroupRowId == groupRowId) {
+            // already visited
+            return;
+        }
+        lastGroupRowId = groupRowId;
+
+        if (over != null) {
+            if (!select.isGroupQuery()) {
+                over.updateAggregate(session, true);
+            }
+        }
+        if (filterCondition != null) {
+            if (!filterCondition.getBooleanValue(session)) {
+                return;
+            }
+        }
+        updateAggregate(session, groupData);
+    }
+
+    /**
+     * Updates an aggregate value.
+     *
+     * @param session
+     *            the session
+     * @param groupData
+     *            group data from the select
+     */
+    protected abstract void updateAggregate(Session session, SelectGroups groupData);
+
+    /**
+     * Invoked when processing group stage of grouped window queries to update
+     * arguments of this aggregate.
+     *
+     * @param session
+     *            the session
+     */
+    protected abstract void updateGroupAggregates(Session session);
 
     protected StringBuilder appendTailConditions(StringBuilder builder) {
         if (filterCondition != null) {
